requirements.txt
flask==3.0.2
mysqlclient
flask-mysqldb
mysql-connector-python
flask_wtf